refactor: auth, user 분리 , 문서 정리 (#82)
