+++ backend/src/docs/asciidoc/post.adoc
@@ -70,3 +70,50 @@ include::{snippets}/post-getList/path-parameters.adoc[]
 성공 응답 HTTP
 include::{snippets}/post-getList/http-response.adoc[]
 
+=== 게시물 참여 신청하기
+
+`DELETE /api/post/{postId}/date/{postDoDateId}/apply`
+
+요청 HTTP
+include::{snippets}/post-apply/http-request.adoc[]
+include::{snippets}/post-apply/path-parameters.adoc[]
+
+성공 응답 HTTP
+include::{snippets}/post-apply/http-response.adoc[]
+
+=== 게시물 참여 취소하기
+
+`POST /api/post/date/{postDoDateId}/apply/cancel`
+
+요청 HTTP
+include::{snippets}/post-apply-cancel/http-request.adoc[]
+include::{snippets}/post-apply-cancel/path-parameters.adoc[]
+
+성공 응답 HTTP
+include::{snippets}/post-apply-cancel/http-response.adoc[]
+
+=== 게시물 즐겨찾기에 추가 or 삭제하기
+
+`POST /api/post/{postId}/like`
+
+요청 HTTP
+include::{snippets}/post-like/http-request.adoc[]
+include::{snippets}/post-like/path-parameters.adoc[]
+
+성공 응답 HTTP
+include::{snippets}/post-like/http-response.adoc[]
+
+=== 게시물 후기 작성하기
+
+`POST /api/post/{postId}/review`
+
+요청 HTTP
+include::{snippets}/post-review-create/http-request.adoc[]
+include::{snippets}/post-review-create/path-parameters.adoc[]
+include::{snippets}/post-review-create/request-fields.adoc[]
+
+성공 응답 HTTP
+include::{snippets}/post-review-create/http-response.adoc[]
+include::{snippets}/post-review-create/response-fields.adoc[]
+
+
