+++ backend/src/docs/asciidoc/user.adoc
@@ -11,66 +11,6 @@
 ---
 == 기본 API
 
-=== 카카오,구글 로그인
-
-`Redirect /oauth2/authorization/kakao`
-
-`Redirect /oauth2/authorization/google`
-
-성공 응답 HTTP
-
-[source]
-----
-http://localhost:3000/callback/userId/nickname/accessToken/schoolId/signupCheck 으로 redirect
-----
-
-[cols=2*]
-|===
-|userId
-|유저 식별자
-|nickname
-|닉네임 회원정보를 입력안했다면 null
-|accessToken
-|accessToken
-|schoolId
-|관심학교 default schooId 회원정보를 입력하지 않았다면 null
-|signupCheck
-|회원가입을 한 유저라면 true 안한 유저라면 false
-|===
-
-=== 회원가입(회원 정보 입력)
-
-`POST /api/user/signup`
-
-요청 HTTP
-include::{snippets}/user-signup/request-parameters.adoc[]
-include::{snippets}/user-signup/request-parts.adoc[]
-
-성공 응답 HTTP
-include::{snippets}/user-signup/http-response.adoc[]
-include::{snippets}/user-signup/response-fields.adoc[]
-
-=== refresh token으로 access token 재발급
-
-`POST /api/user/refreshtoken`
-
-요청 HTTP
-include::{snippets}/user-refresh-token/http-request.adoc[]
-
-성공 응답 HTTP
-include::{snippets}/user-refresh-token/http-response.adoc[]
-include::{snippets}/user-refresh-token/response-fields.adoc[]
-
-=== 로그아웃
-
-`POST /api/user/logout`
-
-요청 HTTP
-include::{snippets}/user-logout/http-request.adoc[]
-
-성공 응답 HTTP
-include::{snippets}/user-logout/http-response.adoc[]
-
 === 내 정보 조회
 
 `GET /api/user/me`
@@ -93,51 +33,3 @@ include::{snippets}/user-update/request-parts.adoc[]
 성공 응답 HTTP
 include::{snippets}/user-update/http-response.adoc[]
 include::{snippets}/user-update/response-fields.adoc[]
-
-== 게시물 관련 API
-
-=== 게시물 참여 신청하기
-
-`DELETE /api/post/{postId}/date/{postDoDateId}/apply`
-
-요청 HTTP
-include::{snippets}/post-apply/http-request.adoc[]
-include::{snippets}/post-apply/path-parameters.adoc[]
-
-성공 응답 HTTP
-include::{snippets}/post-apply/http-response.adoc[]
-
-=== 게시물 참여 취소하기
-
-`POST /api/post/date/{postDoDateId}/apply/cancel`
-
-요청 HTTP
-include::{snippets}/post-apply-cancel/http-request.adoc[]
-include::{snippets}/post-apply-cancel/path-parameters.adoc[]
-
-성공 응답 HTTP
-include::{snippets}/post-apply-cancel/http-response.adoc[]
-
-=== 게시물 즐겨찾기에 추가 or 삭제하기
-
-`POST /api/post/{postId}/like`
-
-요청 HTTP
-include::{snippets}/post-like/http-request.adoc[]
-include::{snippets}/post-like/path-parameters.adoc[]
-
-성공 응답 HTTP
-include::{snippets}/post-like/http-response.adoc[]
-
-=== 게시물 후기 작성하기
-
-`POST /api/post/{postId}/review`
-
-요청 HTTP
-include::{snippets}/post-review-create/http-request.adoc[]
-include::{snippets}/post-review-create/path-parameters.adoc[]
-include::{snippets}/post-review-create/request-fields.adoc[]
-
-성공 응답 HTTP
-include::{snippets}/post-review-create/http-response.adoc[]
-include::{snippets}/post-review-create/response-fields.adoc[]
