+++ backend/src/main/java/com/minibeit/user/web/UserController.java
@@ -1,17 +1,14 @@
 package com.minibeit.user.web;
 
-import com.minibeit.security.token.RefreshTokenService;
 import com.minibeit.security.userdetails.CurrentUser;
 import com.minibeit.security.userdetails.CustomUserDetails;
 import com.minibeit.user.dto.UserRequest;
 import com.minibeit.user.dto.UserResponse;
 import com.minibeit.user.service.UserService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 
 @RestController
@@ -19,43 +16,10 @@ import java.util.List;
 @RequestMapping(path = "/api/user")
 public class UserController {
     private final UserService userService;
-    private final RefreshTokenService refreshTokenService;
-    private static final String REFRESH_TOKEN = "refresh_token";
-
-    //테스트용
-    @PostMapping("/login")
-    public ResponseEntity<UserResponse.Login> login(@RequestBody UserRequest.Login request, HttpServletResponse response) {
-        UserResponse.Login loginResponse = userService.login(request);
-        createCookie(response, loginResponse.getRefreshToken());
-
-        return ResponseEntity.ok().body(loginResponse);
-    }
-
-    @PostMapping("/signup")
-    public ResponseEntity<UserResponse.CreateOrUpdate> signup(UserRequest.Signup request, @CurrentUser CustomUserDetails customUserDetails) {
-        UserResponse.CreateOrUpdate response = userService.signup(request, customUserDetails.getUser());
-        return ResponseEntity.ok().body(response);
-    }
 
     @PostMapping("/update")
     public ResponseEntity<UserResponse.CreateOrUpdate> update(UserRequest.Update request, @CurrentUser CustomUserDetails customUserDetails) {
         return ResponseEntity.ok().body(userService.update(request, customUserDetails.getUser()));
-    }
-
-    @PostMapping("/refreshtoken")
-    public ResponseEntity<UserResponse.Login> refreshToken(@CookieValue("refresh_token") String refreshToken, HttpServletResponse response) {
-        UserResponse.Login loginResponse = refreshTokenService.createAccessToken(refreshToken);
-        createCookie(response, loginResponse.getRefreshToken());
-
-        return ResponseEntity.ok().body(loginResponse);
-    }
-
-    @PostMapping("/logout")
-    public ResponseEntity<Void> logout(@CurrentUser CustomUserDetails customUserDetails, HttpServletResponse response) {
-        userService.logout(customUserDetails.getUser());
-        deleteCookie(response);
-
-        return ResponseEntity.ok().build();
     }
 
     @GetMapping("/me")
@@ -66,28 +30,8 @@ public class UserController {
     }
 
     @GetMapping("/list/business/profile/{businessProfileId}")
-    public ResponseEntity<List<UserResponse.IdAndNickname>> getListInBusinessProfile(@PathVariable Long businessProfileId){
+    public ResponseEntity<List<UserResponse.IdAndNickname>> getListInBusinessProfile(@PathVariable Long businessProfileId) {
         List<UserResponse.IdAndNickname> response = userService.getListInBusinessProfile(businessProfileId);
         return ResponseEntity.ok().body(response);
-    }
-
-    private void createCookie(HttpServletResponse response, String refreshToken) {
-        ResponseCookie cookie = ResponseCookie.from(REFRESH_TOKEN, refreshToken)
-                .httpOnly(true)
-                .path("/")
-                .maxAge(14 * 24 * 60 * 60)
-                .build();
-
-        response.addHeader("Set-Cookie", cookie.toString());
-    }
-
-    private void deleteCookie(HttpServletResponse response) {
-        ResponseCookie cookie = ResponseCookie.from(REFRESH_TOKEN, "")
-                .httpOnly(true)
-                .path("/")
-                .maxAge(0)
-                .build();
-
-        response.addHeader("Set-Cookie", cookie.toString());
     }
 }
