+++ backend/src/test/java/com/minibeit/user/web/UserControllerTest.java
@@ -42,10 +42,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 class UserControllerTest extends MvcTest {
     @MockBean
     private UserService userService;
-    @MockBean
-    private TokenProvider tokenProvider;
-    @MockBean
-    private RefreshTokenService refreshTokenService;
 
     private User user;
 
@@ -62,52 +58,6 @@ class UserControllerTest extends MvcTest {
                 .avatar(File.builder().id(1L).url("profile image url").build())
                 .school(School.builder().id(1L).name("고려대학교").build())
                 .build();
-    }
-
-    @Test
-    @DisplayName("회원가입 문서화")
-    public void signup() throws Exception {
-        InputStream is = new ClassPathResource("mock/images/enjoy.png").getInputStream();
-        MockMultipartFile avatar = new MockMultipartFile("avatar", "avatar.jpg", "image/jpg", is.readAllBytes());
-
-        UserResponse.CreateOrUpdate response = UserResponse.CreateOrUpdate.builder().id(1L).nickname("동그라미").schoolId(1L).build();
-
-        given(userService.signup(any(), any())).willReturn(response);
-
-        ResultActions results = mvc.perform(
-                multipart("/api/user/signup")
-                        .file(avatar)
-                        .param("name", "실명")
-                        .param("nickname", "동그라미")
-                        .param("gender", "MALE")
-                        .param("phoneNum", "010-1234-5678")
-                        .param("job", "대학생")
-                        .param("age", "23")
-                        .param("schoolId", "1")
-                        .contentType(MediaType.MULTIPART_FORM_DATA)
-                        .characterEncoding("UTF-8")
-        );
-
-        results.andExpect(status().isOk())
-                .andDo(document("user-signup",
-                        requestParameters(
-                                parameterWithName("name").description("실명"),
-                                parameterWithName("nickname").description("닉네임"),
-                                parameterWithName("gender").description("성별(MALE or FEMALE)"),
-                                parameterWithName("phoneNum").description("전화번호"),
-                                parameterWithName("job").description("직업"),
-                                parameterWithName("age").description("나이"),
-                                parameterWithName("schoolId").description("관심있는 학교 식별자")
-                        ),
-                        requestParts(
-                                partWithName("avatar").description("사용자 프로필 이미지")
-                        ),
-                        responseFields(
-                                fieldWithPath("id").type(JsonFieldType.NUMBER).description("회원가입한 유저 식별자"),
-                                fieldWithPath("nickname").type(JsonFieldType.STRING).description("회원가입한 유저 닉네임"),
-                                fieldWithPath("schoolId").type(JsonFieldType.NUMBER).description("관심 학교 식별자")
-                        )
-                ));
     }
 
     @Test
@@ -134,43 +84,6 @@ class UserControllerTest extends MvcTest {
                                 fieldWithPath("avatar").type(JsonFieldType.STRING).description("프로필 이미지 url(프로필 이미지가 없다면 null)")
                         )
                 ));
-    }
-
-    @Test
-    @DisplayName("refresh token 문서화")
-    public void refreshToken() throws Exception {
-        UserResponse.Login response = UserResponse.Login.build(1L, "테스터",
-                Token.builder().token("accessToken").expiredAt(LocalDateTime.now()).build(),
-                Token.builder().token("refreshToken").expiredAt(LocalDateTime.now().plusDays(10)).build());
-
-        Cookie cookie = new Cookie("refresh_token", "refreshToken");
-        cookie.setHttpOnly(true);
-        cookie.setMaxAge(14 * 24 * 60 * 60);
-
-        given(refreshTokenService.createAccessToken(any())).willReturn(response);
-        given(tokenProvider.isValidToken(any())).willReturn(true);
-
-        ResultActions results = mvc.perform(post("/api/user/refreshtoken").cookie(cookie));
-
-        results.andExpect(status().isOk())
-                .andDo(print())
-                .andDo(document("user-refresh-token",
-                        responseFields(
-                                fieldWithPath("id").type(JsonFieldType.NUMBER).description("로그인한 유저 식별자"),
-                                fieldWithPath("name").type(JsonFieldType.STRING).description("로그인한 유저 이름"),
-                                fieldWithPath("accessToken").type(JsonFieldType.STRING).description("accessToken")
-                        )
-                ));
-    }
-
-    @Test
-    @DisplayName("logout 문서화")
-    public void logout() throws Exception {
-        ResultActions results = mvc.perform(post("/api/user/logout"));
-
-        results.andExpect(status().isOk())
-                .andDo(print())
-                .andDo(document("user-logout"));
     }
 
     @Test
